keep base API version unaffected by sort.
